Extract run_doctor() from main.rs into doctor.rs module
Moves the ~95-line --doctor diagnostic command into its own module,
reducing main.rs from 2032 to 1940 lines. Ratchet ceiling updated.
